Working through them
