+++ ui/ui-enterprise/pages/model/[modelId]/source/[sourceId]/property/[propertyId]/runs.tsx
@@ -1,6 +1,5 @@
 import Head from "next/head";
 import RunsList from "@grouparoo/ui-components/components/runs/List";
-import { UseApi } from "@grouparoo/ui-components/hooks/useApi";
 import PageHeader from "@grouparoo/ui-components/components/PageHeader";
 import StateBadge from "@grouparoo/ui-components/components/badges/StateBadge";
 import LockedBadge from "@grouparoo/ui-components/components/badges/LockedBadge";
@@ -8,6 +7,7 @@ import PropertyTabs from "@grouparoo/ui-components/components/tabs/Property";
 import { Actions, Models } from "@grouparoo/ui-components/utils/apiData";
 import ModelBadge from "@grouparoo/ui-components/components/badges/ModelBadge";
 import { NextPageContext } from "next";
+import { generateClient } from "@grouparoo/ui-components/client/client";
 
 export default function Page(props) {
   const {
@@ -52,7 +52,7 @@ export default function Page(props) {
 
 Page.getInitialProps = async (ctx: NextPageContext) => {
   const { propertyId, modelId } = ctx.query;
-  const { client } = useApi();
+  const client = generateClient(ctx);
   const { model } = await client.request<Actions.ModelView>(
     "get",
     `/model/${modelId}`
